refactor(usecase): consolidate translate*Err helpers into validators.go
